Added methods:
public Matrix getStdBasisVectorThree(),
public Matrix getStdBasisVectorFour(),
public Matrix getCombBasisVectorThree(),
public Matrix getCombBasisVectorFour().

These methods return polynomial vectors in either the standard basis or the combinatorial basis for dice rolls with two primes or three primes.  Note that as of now, they only work for dice rolls which behave polynomially by the third roll.  So, for example, if a dice has an sides {1, a, b, a^5}, this method will not correctly generate the polynomial describing the growth of unique products of dice rolls.